Allow saving the vault
Here I introduced a workaround for the vault saving:
since I need to have the password in order to both encrypt
and decrypt data, and I need to encrypt the vault before
saving so that it doesn't ask for a password every time
a user wants to save their vault, I needed a more or less secure way of keeping the part of the key responsible
for the encryption of actual data, thus I added the
ProtectedKey property, which might introduce more
vulnerabilities in the program, but will make saving
the vault much easier and faster.

Should think of a way to improve it and the security
in future updates however.