Do pre-release changes
Some of the changes are that 32-bit version is no
longer preferred and the project is now not signed
with a certificate so it should theoretically be
buildable by other people who want to fork it.
Releases will still be signed with private
sertificates (hopefully).
Also fixed a but where dialoghost was missing in
the OptionsView.